Remove thin wall in TC-V6-DOCK-BRACKET-HEMERA-ADAPTER
I noticed the TC-V6-DOCK-BRACKET-HEMERA-ADAPTER.stl featured a thin wall that could prevent the assembly of the components. I thought it was just an error in the STL but it's actually present in the underlying geometry.

I cannot imagine why this feature is present and it is not mentioned in chapter 03b of the assembly.

I have removed the wall in both the STP and the STL files.